# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is the API documentation for the Pismo Assessment project, providing endpoints for account and transaction management.",
        "title": "PismoAssessment API",
        "contact": {
            "name": "Kaustubh Agarwal",
            "email": "kaustubh.agarrwal@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/v1/accounts": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Create a new account",
                "parameters": [
                    {
                        "description": "Account information",
                        "name": "account",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Account"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful account creation",
                        "schema": {
                            "$ref": "#/definitions/models.Account"
                        }
                    },
                    "400": {
                        "description": "Invalid input data",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/accounts/{id}": {
            "get": {
                "description": "Fetches the account details for the specified account ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Get account information by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Account ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "$ref": "#/definitions/models.Account"
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplied",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Account not found",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/transactions": {
            "post": {
                "description": "Initiates a transaction for a specified account, operation type, and amount.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Create a new transaction",
                "parameters": [
                    {
                        "description": "Transaction information",
                        "name": "transaction",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Transaction"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful transaction creation",
                        "schema": {
                            "$ref": "#/definitions/models.Transaction"
                        }
                    },
                    "400": {
                        "description": "Invalid input data",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "errors.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "The HTTP status code of the error.\nexample: 404",
                    "type": "integer"
                },
                "message": {
                    "description": "A detailed error message.\nexample: \"Account not found\"",
                    "type": "string"
                }
            }
        },
        "models.Account": {
            "description": "Account model for managing account information.",
            "type": "object",
            "required": [
                "document_number"
            ],
            "properties": {
                "account_id": {
                    "description": "AccountID is the unique identifier for the account.\n@example 1",
                    "type": "integer"
                },
                "document_number": {
                    "description": "DocumentNumber is the document number associated with the account.\n@example \"12345678900\"\n@validate required,len=11",
                    "type": "string"
                }
            }
        },
        "models.Transaction": {
            "description": "Transaction model for managing transaction information.",
            "type": "object",
            "required": [
                "account_id",
                "amount",
                "operation_type_id"
            ],
            "properties": {
                "account_id": {
                    "description": "AccountID is the identifier for the associated account.\n@example 1",
                    "type": "integer"
                },
                "amount": {
                    "description": "Amount is the amount involved in the transaction.\n@example 100.50",
                    "type": "number"
                },
                "event_date": {
                    "description": "EventDate is the date of the transaction event.\n@example \"2024-10-22T15:04:05Z\"",
                    "type": "string"
                },
                "operation_type_id": {
                    "description": "OperationTypeID is the identifier for the type of operation performed.\n@example 2",
                    "type": "integer"
                },
                "transaction_id": {
                    "description": "TransactionID is the unique identifier for the transaction.\n@example 1",
                    "type": "integer"
                }
            }
        }
    }
}